+++ backend/settings/config.go
@@ -8,6 +8,7 @@ import (
 	"strings"
 
 	"github.com/goccy/go-yaml"
+	"github.com/gtsteffaniak/filebrowser/backend/logger"
 	"github.com/gtsteffaniak/filebrowser/backend/users"
 	"github.com/gtsteffaniak/filebrowser/backend/version"
 )
@@ -15,29 +16,32 @@ import (
 var Config Settings
 
 func Initialize(configFile string) {
-	yamlData := loadConfigFile(configFile)
-	Config = setDefaults()
-	err := yaml.Unmarshal(yamlData, &Config)
+	yamlData, err := loadConfigFile(configFile)
 	if err != nil {
-		log.Fatalf("Error unmarshaling YAML data: %v", err)
+		logger.Warning(fmt.Sprintf("Could not load config file '%v', using default settings: %v", configFile, err))
+	}
+	Config = setDefaults()
+	err = yaml.Unmarshal(yamlData, &Config)
+	if err != nil {
+		logger.Fatal(fmt.Sprintf("Error unmarshaling YAML data: %v", err))
 	}
 	Config.UserDefaults.Perm = Config.UserDefaults.Permissions
 	// Convert relative path to absolute path
 	if len(Config.Server.Sources) > 0 {
 		// TODO allow multipe sources not named default
 		for _, source := range Config.Server.Sources {
-			realPath, err := filepath.Abs(source.Path)
-			if err != nil {
-				log.Fatalf("Error getting source path: %v", err)
+			realPath, err2 := filepath.Abs(source.Path)
+			if err2 != nil {
+				logger.Fatal(fmt.Sprintf("Error getting source path: %v", err2))
 			}
 			source.Path = realPath
 			source.Name = "default"                   // Modify the local copy of the map value
 			Config.Server.Sources["default"] = source // Assign the modified value back to the map
 		}
 	} else {
-		realPath, err := filepath.Abs(Config.Server.Root)
-		if err != nil {
-			log.Fatalf("Error getting source path: %v", err)
+		realPath, err2 := filepath.Abs(Config.Server.Root)
+		if err2 != nil {
+			logger.Fatal(fmt.Sprintf("Error getting source path: %v", err2))
 		}
 		Config.Server.Sources = map[string]Source{
 			"default": {
@@ -67,28 +71,46 @@ func Initialize(configFile string) {
 			Url:  "https://github.com/gtsteffaniak/filebrowser/wiki",
 		})
 	}
+	if len(Config.Server.Logging) == 0 {
+		Config.Server.Logging = []LogConfig{
+			{
+				Output: "stdout",
+			},
+		}
+	}
+	for _, logConfig := range Config.Server.Logging {
+		err = logger.SetupLogger(
+			logConfig.Output,
+			logConfig.Levels,
+			logConfig.ApiLevels,
+			logConfig.NoColors,
+		)
+		if err != nil {
+			log.Println("[ERROR] Failed to set up logger:", err)
+		}
+	}
+
 }
 
-func loadConfigFile(configFile string) []byte {
+func loadConfigFile(configFile string) ([]byte, error) {
 	// Open and read the YAML file
 	yamlFile, err := os.Open(configFile)
 	if err != nil {
-		log.Println(err)
-		os.Exit(1)
+		return nil, err
 	}
 	defer yamlFile.Close()
 
 	stat, err := yamlFile.Stat()
 	if err != nil {
-		log.Fatalf("error getting file information: %s", err.Error())
+		return nil, err
 	}
 
 	yamlData := make([]byte, stat.Size())
 	_, err = yamlFile.Read(yamlData)
 	if err != nil {
-		log.Fatalf("Error reading YAML data: %v", err)
+		return nil, err
 	}
-	return yamlData
+	return yamlData, nil
 }
 
 func setDefaults() Settings {
@@ -101,7 +123,6 @@ func setDefaults() Settings {
 			NumImageProcessors: 4,
 			BaseURL:            "",
 			Database:           "database.db",
-			Log:                "stdout",
 			Root:               ".",
 		},
 		Auth: Auth{